some shop functionality/better menu navigation
battling is broken - something to do with pets not getting properly assigned to their teams in the randomization process. I'm working on it.
Also, a big thing that needs to be added is setting the team back to a "healthy" state once the battle is over, so it can be edited. This has to do with objects/references and will take a moment to figure out.